backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
anthropic>=0.39.0
python-dotenv>=1.0.0
slowapi>=0.1.9
